fix: correct the misuse of source Tree
The past attempt to override the recipe has not been effective. In
reality, it is necessary to generate a JSON from the recipe to override
the source code hash and dependencies.

To make the recipe work properly, I would need to maintain a separate
script, which is too cumbersome. Therefore, I am currently using the
patches instead.

Signed-off-by: Burgess Chang <bsc@brsvh.org>